refactor. update plt update data
